+++ hm.js
@@ -10,6 +10,7 @@ import { HMItemFactory } from './modules/item/item-factory.js';
 import { HMItemSheet } from './modules/item/item-sheet.js';
 import { HMArmorItemSheet } from './modules/item/armor-item-sheet.js';
 import { HMClassItemSheet } from './modules/item/class-item-sheet.js';
+import { HMProficiencyItemSheet } from './modules/item/proficiency-item-sheet.js';
 import { HMRaceItemSheet } from './modules/item/race-item-sheet.js';
 import { HMSpellItemSheet } from './modules/item/spell-item-sheet.js';
 import { HMWeaponItemSheet } from './modules/item/weapon-item-sheet.js';
@@ -32,6 +33,7 @@ function registerSheets() {
     Items.registerSheet('hackmaster', HMItemSheet, {makeDefault: true});
     Items.registerSheet('hackmaster', HMArmorItemSheet, {types: ['armor'], makeDefault: true});
     Items.registerSheet('hackmaster', HMClassItemSheet, {types: ['cclass'], makeDefault: true});
+    Items.registerSheet('hackmaster', HMProficiencyItemSheet, {types: ['proficiency'], makeDefault: true});
     Items.registerSheet('hackmaster', HMRaceItemSheet, {types: ['race'], makeDefault: true});
     Items.registerSheet('hackmaster', HMSpellItemSheet, {types: ['spell'], makeDefault: true});
     Items.registerSheet('hackmaster', HMWeaponItemSheet, {types: ['weapon'], makeDefault: true});