+++ modules/item/proficiency-item-sheet.js
@@ -1,0 +1,13 @@
+import { HMItemSheet } from './item-sheet.js';
+
+export class HMProficiencyItemSheet extends HMItemSheet {
+    /** @override */
+    static get defaultOptions() {
+        return mergeObject(super.defaultOptions, {
+            classes: ['hackmaster', 'sheet', 'item'],
+            width: 480,
+            height: 280,
+            tabs: [{ navSelector: '.sheet-tabs', contentSelector: '.sheet-body', initial: 'attributes' }],
+        });
+    }
+}